Add visual studio boilerplate to compile for Windows
Checking in a bunch of binaries is a bit gross but oh well it's a
prototype.

Tested and working with Visual Studio 2017, building a 32 bit binary.

libuv was built from master, 620be96fc4c7645b606348933c67dbc201f53d28.

libbson was built from 1.9.2. The binaries in libbson-bin are loaded by
the application because VS is configured to set the PATH for the
executable, and PATH is the final place apps search for DLLs. Statically
linking libbson would be nicer but I ran into problems with (I think)
multiple versions of the same function being defined.